+++ tests/src/rules/no-unresolved.js
@@ -451,10 +451,19 @@ context('TypeScript', () => {
           code: 'import type { JSONSchema7Type } from "@types/json-schema";',
           parser,
         }),
+        test({
+          code: 'export type { JSONSchema7Type } from "@types/json-schema";',
+          parser,
+        }),
       ],
       invalid: [
         test({
           code: 'import { JSONSchema7Type } from "@types/json-schema";',
+          errors: [ "Unable to resolve path to module '@types/json-schema'." ],
+          parser,
+        }),
+        test({
+          code: 'export { JSONSchema7Type } from "@types/json-schema";',
           errors: [ "Unable to resolve path to module '@types/json-schema'." ],
           parser,
         }),